Beginning Weight matrix and getting cols/rows
Added integers to store the number of columns and rows for each training
sample (so I can reformat the Integer list as a 2D array for each
sample). Began writing the weight matrix, but it doesn't work correctly
as of yet.
Also added some catches if user inputs a blank file or an incorrectly
formatted file.