AHGHGHGH Figured out bug in first training pair
Took me a while, but I figured out that in the first training pair, it
wasn't reading a full 10 indices for each line, some only had 9 (when
the last index held a space) and others had 10. Fixed it by adding the
space to the training file. Now to double check the transpose and
multiplication arrays.